Add table to index.html, loop through array
Add a table to index.html with four rows which will be the headers.
In js file, add a function that takes in an argument, which will be the
book object, and this function pushes the book object to the library
array. After that, Select the table tag in html and then add a row for
every book in the array, and the text content in this row is the object
as a string