Incorporate frontend display of library
Use display: grid to display two rows, one for the header and a larger
one for the body. Add in divs inside of shelf that have a width and
height. Then add in styling to these divs which will be the "cards".
Then in javascript, have it so we add in a card for every book that
is in the array.