Add in comments for readability, button and form
Add in comments in javascript to be able to easily go through code and
figure out what is happening. Add in a button with the same styling as
the book divs that will eventually on click bring up a form. Start to
make a form and eventually style and then hide it