Add styling to form
Use position fixed and left and top and translate properties to center
div with white background in the center of the screen at all times.
Have a set width and make height auto. Display none at the start.
Js file will then display block when button is clicked. Add styling to
inputs and labels and select. #myForm is the form and will have padding
and is in display flex and flex-direction column so all labels will be
followed by its input or drop down in the next line down