Add new book and more styling and interactability
Have buttons inside the form be just a button type and not a submit type
which is the default. Then select close button in js file and then have
it so on click it will hide the form again. Add another book constructor
into our array. Add an image file input into the form. Add a hover
effect into our books and add button by going into CSS and changing the
background color on :hover.