Style add book button once clicked
Have the background color of the plus button stay as grey after clicking
by having focus as well as hover in the CSS selection. Then when the
button is clicked, add another class called normalCursor that makes
the cursor go back to default so it doesn't feel like a button anymore.
Remove this class when close is clicked and eventually when add is
clicked