Reset form and add in image to the new card
After getting all the contents from formData, use .reset() on the form
to clear the content and also not reload the page. Remove the img.src
HTML from the innerHTML call. Add in a FileReader that reads the file
and creates an img element with the contents being the image link and
then appends this img element child to the new card thats being
developed on click.