Display image on file input after add button
When add button is clicked, have a variable imageName that gets the
file input and then have a new FileReader. This file reader will run
readAsDataUrl on this imageName. Then on reader.onload create an img
element and the .src of this will be the result. Then append this img
onto newCard1