Move picture and new div to top of shelf
Restructure how the image gets into the html by adding it to
innerHTML rather than creating a new element. Then directly add
e.target.result on the innerHTML. Add a span into the shelf so we can
use insertBefore and change styling to have divs and buttons be on
margin-left rather than gaps for flexbox in shelf to avoid any spacing
from the added span. Then add in a new class for addBookButton so that
it will remain grey until the form is not displayed any longer.