Add in read checkbox and revise adding new book
Revise adding new book to not run the updateShelf() function but just
adding HTML based on current inputs from form listener like previous.
Then add in checkbox in both old and new HTML library cards and add
a class that receives a green background if finished reading is clicked
in the form or if it received yes.