Add required attribute, unique id for checkbox
Add styling on border for delete button in css file. i is a random
iterator that increases after looping through the array of books.
It will also increase after adding a book using the plus button. So we
use this iterator and assign it to id attribute of each checkbox so each
will be unique input id. Then add a required attribute in the form for
title author and number of pages